// src/api/generated/model/skillROI.ts
/**
 * Generated by orval v7.0.0 🍺
 * Do not edit manually.
 * MHWs Talisman Vision API
 * Backend API and Vision inference engine for MHWs Talisman Vision Module.
 * OpenAPI spec version: 0.1.0
 */
import type { RelativeRect } from './relativeRect';

export interface SkillROI {
  id: number;
  level: RelativeRect;
  name: RelativeRect;
}