// src/api/generated/model/normalizationPoints.ts
/**
 * Generated by orval v7.0.0 🍺
 * Do not edit manually.
 * MHWs Talisman Vision API
 * Backend API and Vision inference engine for MHWs Talisman Vision Module.
 * OpenAPI spec version: 0.1.0
 */
import type { Point } from './point';

export interface NormalizationPoints {
  bg_point: Point;
  frame_point: Point;
}